Make letsencrypt handle config change more reliably
When config.sh is changed (e.g. dev to prod), we usually need to
regenerate the certificates. dehydrate checks the certificate age and
chickens out if there is more than 7 days left on it, this change
deletes the certificate directory and re-runs the wrapper.

I've also moved the "retain config" task to the end so that it happens
after the handlers. This should mean that we don't get a changed task
when running immediately after the initial deploy.

Signed-off-by: Ian Allison <iana@pims.math.ca>