Google and Github authenticator mods, shibboleth-sp and html changes
This commit goes with the syzygy branch of the syzygyauthenticator to
allow google and github oauth targets. It also includes the old
shibboleth-sp role for compatibility with existing syzygy hosts. Ideally
the OAuth stuff will be mixed in with the ssp-idp stuff from callysto,
but for the moment we'll just handle it as before for continuity.

Signed-off-by: Ian Allison <iana@pims.math.ca>