Minor niggles remove packages and fix up stats apache role
I forgot to remove the python36u packages ahead of trying to install the
system python36 ones this chokes yum. I also forgot that the role
variables for apache wouldn't be available when running the stats-client
play. For the moment I've shoe-horned some extra variables with sensible
defaults but we might want to handle this more systematically in future.

Signed-off-by: Ian Allison <iana@pims.math.ca>