ZOL releases trail redhat by a but causing 404s
The ZFS on Linux project normally makes repo releases whenever there is
a minor update to CentOS/Almalinux. Unfortunately there is sometimes a
lag between these releases. Normally, we detect the distro version to
know what repo to ask for, but if it isn't yet available, this change
allows us to explicitly set zfs_distro_version in host_vars as needed.

Signed-off-by: Ian Allison <iana@pims.math.ca>